STYLE: put main in new klaplace.cxx source file for further extensions of klaplace
